requirements.txt
amqp==2.4.2
billiard==3.5.0.5
celery==4.2.1
Django==2.1.5
django-filter==2.1.0
django-rest-framework==0.1.0
djangorestframework==3.9.2
kombu==4.4.0
Markdown==3.0.1
psycopg2==2.7.7
pytz==2018.9
redis==3.2.0
vine==1.2.0
